requirements.txt
Django==3.1.3
django-cacheops==5.1
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-mptt==0.11.0
django-pglocks==1.0.4
django-prometheus==2.1.0
django-rq==2.4.0
django-tables2==2.3.3
django-taggit==1.3.0
django-timezone-field==4.0
djangorestframework==3.12.2
drf-yasg[validation]==1.20.0
gunicorn==20.0.4
Jinja2==2.11.2
Markdown==3.3.3
netaddr==0.8.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pycryptodome==3.9.9
PyYAML==5.3.1
svgwrite==1.4

